feat - 구조 변경

+++ src/main/java/com/student/domain/controller/StudentController.java
@@ -1,22 +1,33 @@
 package com.student.domain.controller;
 
 
+import com.student.Error.ApiResponseMessage;
 import com.student.Error.ErrorsResource;
 import com.student.domain.Student;
+import com.student.domain.StudentResource;
 import com.student.domain.StudentValidator;
+import com.student.domain.dto.SearchDto;
 import com.student.domain.dto.StudentDto;
 import com.student.service.StudentService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PagedResourcesAssembler;
+import org.springframework.hateoas.Link;
 import org.springframework.hateoas.MediaTypes;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.net.URI;
+import java.util.Optional;
+
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 
 @Slf4j
 @Controller
@@ -25,6 +36,7 @@ import javax.validation.Valid;
 public class StudentController {
     private final StudentValidator studentValidator;
     private final StudentService studentService;
+    private final ModelMapper modelMapper;
 
     /**
      * 학생 생성 Controller
@@ -34,7 +46,8 @@ public class StudentController {
      * @return
      */
     @PostMapping
-    public ResponseEntity createStudent(@RequestBody @Valid StudentDto studentDto, Errors errors) {
+    public @ResponseBody
+    ResponseEntity createStudent(@Valid StudentDto studentDto, Errors errors) {
         if (errors.hasErrors()) {
             return badRequest(errors);
         }
@@ -43,17 +56,44 @@ public class StudentController {
             return badRequest(errors);
         }
 
-        return studentService.createStudent(studentDto);
+        Student searchStudent = studentService.searchStudent(studentDto);
+
+        if (searchStudent != null) {
+            ApiResponseMessage message = new ApiResponseMessage("Fail", "overlapData", "", "");
+            return new ResponseEntity<ApiResponseMessage>(message, HttpStatus.BAD_REQUEST);
+        }
+
+        Student student = studentService.createStudent(studentDto);
+
+        var selfLinkBuilder = linkTo(StudentController.class).slash(student.getId());
+        URI createdUri = selfLinkBuilder.toUri();
+        StudentResource studentResource = new StudentResource(student);
+        studentResource.add(linkTo(StudentController.class).withRel("query-students"));
+        studentResource.add(selfLinkBuilder.withRel("update-student"));
+        studentResource.add(Link.of("/docs/index.html#resources-student-create").withRel("profile"));
+
+        return ResponseEntity.created(createdUri).body(studentResource);
     }
 
     /**
      * 학생 한명 조회
+     *
      * @param id
      * @return
      */
     @GetMapping("/{id}")
-    public ResponseEntity getStudent(@PathVariable Integer id) {
-        return studentService.getStudent(id);
+    public @ResponseBody
+    ResponseEntity getStudent(@PathVariable Integer id) {
+        Student student = studentService.getStudent(id);
+
+        if (student == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        StudentResource studentResource = new StudentResource(student);
+        studentResource.add(Link.of("/docs/index.html#resources-student-get").withRel("profile"));
+
+        return ResponseEntity.ok(studentResource);
     }
 
     /**
@@ -61,25 +101,58 @@ public class StudentController {
      *
      * @param pageable
      * @param assembler
+     * @param searchDto
      * @return
      */
     @GetMapping
-    public ResponseEntity queryStudents(Pageable pageable, PagedResourcesAssembler<Student> assembler) {
-        return studentService.queryStudent(pageable, assembler);
+    public @ResponseBody
+    ResponseEntity queryStudents(@Valid SearchDto searchDto, Pageable pageable, PagedResourcesAssembler<Student> assembler) {
+
+        Page<Student> studentPages = studentService.queryStudent(pageable, searchDto);
+
+        var pageResource = assembler.toModel(studentPages, e -> new StudentResource(e));
+        pageResource.add(Link.of("/docs/index.html#resources-events-list").withRel("profile"));
+        pageResource.add(linkTo(StudentController.class).withRel("create-event"));
+
+        return ResponseEntity.ok(pageResource);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity updateStudent(@PathVariable Integer id,
-                                        @RequestBody @Valid StudentDto studentDto, Errors errors) {
+    public @ResponseBody
+    ResponseEntity updateStudent(@PathVariable Integer id,
+                                 @Valid StudentDto studentDto, Errors errors) {
         if (errors.hasErrors()) {
             return badRequest(errors);
         }
-        return studentService.updateStudent(id, studentDto);
+        Student student = studentService.getStudent(id);
+
+        if (student == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        Student existingStudent = studentService.updateStudent(id, studentDto);
+        StudentResource studentResource = new StudentResource(existingStudent);
+        studentResource.add(Link.of("/docs/index.html#resources-student-update").withRel("profile"));
+
+        return ResponseEntity.ok(studentResource);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteStudent(@PathVariable Integer id){
-        return studentService.deleteStudent(id);
+    public @ResponseBody
+    ResponseEntity deleteStudent(@PathVariable Integer id, Errors errors) {
+        if (errors.hasErrors()) {
+            return badRequest(errors);
+        }
+        Student student = studentService.getStudent(id);
+        if (student == null) {
+            return ResponseEntity.notFound().build();
+        }
+
+        Student deleteStudent = studentService.deleteStudent(id);
+        StudentResource studentResource = new StudentResource(deleteStudent);
+        studentResource.add(Link.of("/docs/index.html#resources-student-delete").withRel("profile"));
+
+        return ResponseEntity.ok(studentResource);
     }
